Fix data loss when messages sent during conf
Problem
========

Suppose we have:

```
  ModuleA --q1--> Q2N (ipm::Sender) ==network==> N2Q --q2--> ModuleB
```

And both:

* ModuleA is sent "conf" before ModuleB
* ModuleA pushes to q1 during its do_configure() function

Then Q2N will pop the item off its queue and attempt to send
the item _before_ ModuleB connects. The connection is not yet
made, and so the network send() times out and the item is
dropped. Q2N has no way to tell ModuleA that the item has been
dropped, so we get silent data loss.

Solution/Workaround
===================

We add the concept of "control" messages to
NetworkObjectSender/NetworkObjectReceiver, in addition to regular
messages that contain a serialized object. Then Q2N sends a single
"control" message before it starts popping items from the input
queue. If this message times out, we throw an exception and so never
get to the point where we're popping items from the input queue.

In practice, this means that the configure command will fail when a
Q2N is configured before its corresponding N2Q. That stops the whole
application. This is kind of ugly, but it's better than silently
dropping data.